Create the basic game logic
Set up the game and the game over view, and create code that allows the
user to shoot bombs at donuts. When a user hits a donut, destroy it and
the bomb, and give them a point. Also, implement an ammo mechanism in
order to prevent the user from shooting too many bombs.